fix: move styled components to a single file
