--- src/components/Calendar/doubleCalendar/styled/calendar.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const StyledCalendar = styled.div`
-    display: flex;
-    flex: 1;
-`;
-
-export default StyledCalendar;
