--- src/components/Calendar/doubleCalendar/styled/container.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const StyledContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-content: start;
-`;
-
-export default StyledContainer;
