--- src/components/Calendar/doubleCalendar/styled/divider.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const StyledDivider = styled.div`
-    display: flex;
-    width: 5rem;
-`;
-
-export default StyledDivider;
