--- src/components/Calendar/doubleCalendar/styled/headerContainer.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const StyledHeaderContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-grow: 1;
-`;
-
-export default StyledHeaderContainer;
