--- src/components/Calendar/doubleCalendar/styled/monthsContainer.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const StyledMonthsContainer = styled.div`
-    display: flex;
-    justify-content: stretch;
-    align-items: center;
-    flex-grow: 1;
-`;
-
-export default StyledMonthsContainer;
